backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Pydantic / settings
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator>=2.0.0

# Auth
bcrypt==4.2.1
python-multipart==0.0.17

# AI
anthropic==0.40.0

# HTTP client (MS Graph, webhooks)
httpx==0.28.1

# Email / IMAP
# imaplib and smtplib are part of stdlib — no extra package needed

# Environment
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0
